native style titlebar
- replace header with a fixed, native app style titlebar
- move search, toolbar into titlebar
- remove local emotion dependency; replace with tailwind where used (2 places)
- abstract various small things like FrontMatter and scroll container
- replace a few usages of evergreen-ui with tailwind classes
