+++ package.json
@@ -55,7 +55,6 @@
     "class-variance-authority": "^0.7.0",
     "date-fns": "^3.3.1",
     "electron": "^28.2.0",
-    "emotion": "^10.0.27",
     "esbuild": "^0.20.0",
     "evergreen-ui": "^7.1.9",
     "klaw": "^3.0.0",
