+++ src/components/icons.tsx
@@ -25,6 +25,7 @@ import {
   ExternalLink,
   Eye,
   FileCode,
+  FilePenLineIcon,
   Folder,
   FolderArchive,
   FolderCheck,
@@ -203,7 +204,7 @@ export const Icons = {
   commentAdd: MessageSquarePlus,
   delete: Trash2,
   dragHandle: GripVertical,
-  editing: Edit2,
+  editing: FilePenLineIcon,
   emoji: Smile,
   externalLink: ExternalLink,
   folder: Folder,
