+++ src/index.css
@@ -26,6 +26,7 @@
     --secondary: 210 40% 96.1%;
     --secondary-foreground: 222.2 47.4% 11.2%;
 
+    /* --accent: 255, 89%, 62%; */
     --accent: 210 40% 96.1%;
     --accent-foreground: 222.2 47.4% 11.2%;
