+++ src/views/documents/search/index.tsx
@@ -35,6 +35,7 @@ const SearchDocuments = (props: Props) => {
 
   return (
     <TagInput
+      className="drag-none"
       flexGrow={1}
       inputProps={{ placeholder: "Search journals" }}
       values={props.store.searchTokens}
