+++ src/views/edit/loading.tsx
@@ -1,8 +1,9 @@
 import React from "react";
 import { observer } from "mobx-react-lite";
-import { Pane, Button, Alert } from "evergreen-ui";
-import { css } from "emotion";
+import { Pane, IconButton, ChevronLeftIcon } from "evergreen-ui";
 import { useNavigate } from "react-router-dom";
+import Titlebar from "../../titlebar/macos";
+import { Separator } from "./editor/components/Separator";
 
 export interface LoadingComponentProps {
   error?: Error | null;
@@ -15,68 +16,27 @@ export const EditLoadingComponent = observer((props: LoadingComponentProps) => {
   const navigate = useNavigate();
 
   return (
-    <Pane>
-      <a onClick={() => navigate("/documents")}>Back</a>
-      <Pane marginTop={24}>
-        <div
-          className={css`
-            display: flex;
-            justify-content: flex-start;
-          `}
+    <>
+      <Titlebar>
+        <IconButton
+          backgroundColor="transparent"
+          border="none"
+          icon={ChevronLeftIcon}
+          className="drag-none"
+          onClick={() => {}}
+          marginRight={8}
         >
-          <div
-            className={css`
-              margin-right: 4px;
-            `}
-          >
-            {placeholderDate}/
-          </div>
-          <span
-            className={css`
-              border-bottom: 1px dotted purple;
-              color: purple;
-              cursor: pointer;
-            `}
-          >
-            Loading...
-          </span>
-        </div>
-        <div
-          className={css`
-            margin-bottom: 16px;
-            margin-top: 16px;
-          `}
-        >
-          <input
-            type="text"
-            name="title"
-            className={css`
-              font-size: 1.5em;
-              border: none;
-              width: 100%;
-              &:focus {
-                outline: none;
-              }
-            `}
-            placeholder=""
-            disabled={true}
-          />
-        </div>
-
-        {/* note: its not actually clear to me whether toJS is necessary here. */}
-        <Pane></Pane>
-        {props.error && (
-          <Alert intent="danger" title="Error loading documents">
-            {JSON.stringify(props.error)}
-          </Alert>
-        )}
-
-        <Pane marginTop={24}>
-          <Button onClick={noop} disabled={true} isLoading={true}>
-            Save
-          </Button>
+          Back to documents
+        </IconButton>
+        <Separator orientation="vertical" />
+      </Titlebar>
+      <Pane padding={50} paddingTop={98} flexGrow={1} display="flex">
+        <Pane flexGrow={1} display="flex" flexDirection="column" width="100%">
+          <Pane flexGrow={1} paddingTop={24}>
+            <h3>Coming soon...</h3>
+          </Pane>
         </Pane>
       </Pane>
-    </Pane>
+    </>
   );
 });
